# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/support-tickets": {
      "post": {
        "operationId": "SupportController_createTicket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create support ticket",
        "tags": [
          "Support"
        ]
      },
      "get": {
        "operationId": "SupportController_getMyTickets",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "RESOLVED",
                "CLOSED"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List my support tickets",
        "tags": [
          "Support"
        ]
      }
    },
    "/support-tickets/{id}": {
      "get": {
        "operationId": "SupportController_getTicketDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get ticket detail",
        "tags": [
          "Support"
        ]
      },
      "patch": {
        "operationId": "SupportController_updateTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update support ticket",
        "tags": [
          "Support"
        ]
      }
    },
    "/support-tickets/{id}/messages": {
      "get": {
        "operationId": "SupportController_getMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for cursor-based pagination",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Load ticket messages",
        "tags": [
          "Support"
        ]
      },
      "post": {
        "operationId": "SupportController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupportMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Send message",
        "tags": [
          "Support"
        ]
      }
    },
    "/admin/support": {
      "get": {
        "operationId": "AdminSupportController_adminGetTickets",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "RESOLVED",
                "CLOSED"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Admin list tickets",
        "tags": [
          "Admin - Support"
        ]
      }
    },
    "/admin/support/{id}/join": {
      "post": {
        "operationId": "AdminSupportController_adminJoinTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Admin join ticket",
        "tags": [
          "Admin - Support"
        ]
      }
    },
    "/admin/support/{id}": {
      "patch": {
        "operationId": "AdminSupportController_adminUpdateTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Admin update ticket",
        "tags": [
          "Admin - Support"
        ]
      }
    },
...
  },
  "info": {
    "title": "RWA Backend API",
    "description": "API documentation for the RWA graduation project - Real World Assets Platform",
    "version": "1.0.0",
    "contact": {
      "name": "RWA Team",
      "url": "",
      "email": "support@rwa.com"
    },
    "license": {
      "name": "Proprietary",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      },
      "accessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "name": "apiKey",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateSupportTicketDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Ticket subject",
            "enum": [
              "Account Issue",
              "Deposit",
              "Withdrawal",
              "Trading Issue",
              "Technical Issue",
              "Verification",
              "Billing",
              "Other"
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000,
            "description": "Initial message content"
          },
          "title": {
            "type": "string",
            "description": "Optional title"
          },
          "category": {
            "type": "string",
            "description": "Optional category"
          },
          "priority": {
            "type": "string",
            "description": "Optional priority (free-form string)"
          }
        },
        "required": [
          "subject",
          "message"
        ]
      },
      "UpdateSupportTicketDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Ticket subject"
          },
          "title": {
            "type": "string",
            "description": "Optional title"
          },
          "category": {
            "type": "string",
            "description": "Optional category"
          },
          "priority": {
            "type": "string",
            "description": "Optional priority (free-form string)"
          },
          "status": {
            "type": "string",
            "description": "Ticket status",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED",
              "CLOSED"
            ]
          }
        }
      },
      "SupportMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000,
            "description": "Message content"
          },
          "messageType": {
            "type": "number",
            "default": 0,
            "description": "Message type. 0 = TEXT, 1 = IMAGE",
            "enum": [
              0,
              1
            ]
          }
        },
        "required": [
          "content"
        ]
      },
...
      "ValidationErrorItem": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Machine-friendly field-level error code."
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "code",
          "t"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string",
            "enum": [
              "internal-error",
              "not-found",
              "bad-request",
              "validation-error",
              "rate-limit-exceeded",
              "duplicate-request",
              "unauthorized",
              "forbidden",
              "token-expired",
              "asset-not-found",
              "invalid-token",
              "token-not-found",
              "session-not-found",
              "session-revoked",
              "session-expired",
              "user-not-found",
              "user-not-active",
              "user-already-exists",
              "invalid-credentials",
              "reference-code-not-found",
              "reference-code-invalid",
              "reference-code-already-set",
              "password-attempt-limit-exceeded",
              "otp-not-found",
              "invalid-otp",
              "invalid-otp-token",
              "mfa-already-enabled",
              "mfa-not-enabled",
              "invalid-mfa-code",
              "challenge-required",
              "challenge-not-found",
              "challenge-expired",
              "challenge-already-used",
              "challenge-invalid-factor",
              "challenge-payload-mismatch",
              "account-not-found",
              "account-not-active",
              "position-not-found",
              "position-already-closed",
              "feedback-not-found",
              "post-not-found",
              "post-unauthorized",
              "comment-not-found",
              "comment-unauthorized",
              "support-ticket-not-found",
              "support-unauthorized",
              "insufficient-balance",
              "transfer-to-same-account",
              "transfer-not-allowed-with-open-position",
              "external-api-error",
              "invalid-quantity",
              "price-not-found"
            ]
          },
          "message": {
            "type": "string",
            "description": "Human-readable error summary for clients/logging."
          },
          "t": {
            "type": "string",
            "format": "date-time"
          },
          "errors": {
            "description": "Detailed error information by field, domain rule, or external provider.",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorItem"
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "traceId": {
            "type": "string",
            "description": "Request trace id for debugging across services."
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request payload or failed domain validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is missing, invalid, revoked, or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed to access or mutate this resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded or too many OTP/password attempts.",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Unexpected server error or upstream dependency failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}